requirements_windows.txt
# Windows client dependencies
nfcpy>=1.0.4
pyscard>=2.0.7
requests>=2.28.0

